service-ocr/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==10.1.0
pillow-heif==0.13.1
pdf2image==1.16.3
httpx==0.25.1